requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.10
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
bitarray==3.8.0
blinker==1.4
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.2
cryptography==3.4.8
cytoolz==1.1.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
exceptiongroup==1.3.1
fastapi==0.136.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib-metadata==4.6.4
iniconfig==2.3.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.7.0
oauthlib==3.2.0
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
propcache==0.4.1
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.1
py_order_utils==0.3.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==9.0.2
python-apt==2.4.0+ubuntu4
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
SecretStorage==3.3.1
six==1.16.0
starlette==1.0.0
tomli==2.3.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unattended-upgrades==0.1
urllib3==2.6.2
uvicorn==0.44.0
wadllib==1.3.6
websocket-client==1.8.0
yarl==1.22.0
zipp==1.0.0
